telegram_bot/requirements.txt
# Telegram Bot
python-telegram-bot==21.0.1

# HTTP Client
httpx==0.27.0

# Redis
redis==5.0.1

# Configuration
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
